requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
opencv-python
numpy
Pillow
pytorch-grad-cam
python-dotenv
httpx
